# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "InterMine-Configurator prototype spec",
    "version": "0.1.0",
    "title": "InterMine Configurator API Specs - working draft",
    "contact": {
      "email": "info@intermine.org",
      "url": "http://www.intermine.org"
    },
    "license": {
      "name": "LGPL 2.1",
      "url": "https://github.com/intermine/intermine/blob/dev/LICENSE"
    }
  },
  "paths": {
...
    "/file/properties/detect": {
      "post": {
        "summary": "Return information about this file",
        "description": "Reads in data file, validates and returns file attributes, plus questions (if needed) for the user.",
        "operationId": "detectFileProperties",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "ID of user who owns this mine",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-556642440000"
            }
          },
          {
            "name": "mineId",
            "in": "query",
            "description": "ID of mine",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "3103a884-ff0b-4e77-9c4f-1354a5b928d5"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfully detected file type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataFileProperties"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/DataFileProperties"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataFile"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/DataFile"
              }
            }
          },
          "description": "File that needs to be identified.",
          "required": true
        }
      }
    },
    "/file/properties/save": {
      "post": {
        "summary": "Save file properties to a given InterMine pre-build config",
        "description": "Following on from `detectFileProperties`, we need to save a set of metadata associated with a file. This can be run multiple times, as the user may upload an unknown number of files.",
        "operationId": "saveFileProperties",
        "parameters": [
          {
            "name": "mineId",
            "in": "query",
            "description": "ID of mine to fetch",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-556642440000"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "description": "ID of user who owns this mine",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "4187cd22-7709-45f0-af33-94b36b6bf561"
            }
          },
          {
            "name": "fileId",
            "in": "query",
            "description": "ID of file to save",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "0bca2cef-2541-49ec-84d9-0acb434a8c3c"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfully saved provided DataFileProperties details"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataFilePropertiesResponse"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/DataFilePropertiesResponse"
              }
            }
          },
          "description": "File that needs to be identified.",
          "required": true
        }
      }
    },
...
  },
  "servers": [{
    "url": "https://wizard.intermine.org/v1"
  },
    {
      "url": "http://wizard.intermine.org/v1"
    }
  ],
  "components": {
    "schemas": {
...
      "Organism": {
        "type": "object",
        "required": [
          "taxonID"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "Homo sapiens"
          },
          "taxonID": {
            "type": "integer",
            "example": "9606"
          }
        }
      },
      "DataFileProperties": {
        "type": "object",
        "description": "All of the config input by the user via the wizard",
        "required": [
          "dataFile"
        ],
        "properties": {
          "dataFile": {
            "allOf": [{
              "$ref": "#/components/schemas/DataFile"
            }]
          },
          "descriptors": {
            "type": "array",
            "description": "Information about the file, e.g. entity count, sequence type",
            "items": {
              "$ref": "#/components/schemas/DataFileDescriptor"
            }
          },
          "filePreview": {
            "allOf": [{
              "$ref": "#/components/schemas/DataFilePreview"
            }]
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataFilePropertiesQuestion"
            }
          }
        },
        "xml": {
          "name": "DataFileProperties"
        }
      },
      "DataFile": {
        "type": "object",
        "description": "Represents the data file",
        "required": [
          "name",
          "fileId",
          "fileFormat",
          "organism"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "humanfile.gff"
          },
          "fileId": {
            "type": "string",
            "format": "uuid",
            "example": "03641b4d-bd7e-402f-803a-7aaf55c17238"
          },
          "fileFormat": {
            "type": "string",
            "enum": [
              "fasta",
              "gff",
              "tab",
              "csv"
            ],
            "example": "fasta"
          },
          "organism": {
            "allOf": [{
              "$ref": "#/components/schemas/Organism"
            }]
          }
        },
        "xml": {
          "name": "DataFile"
        }
      },
      "DataFilePreview": {
        "type": "object",
        "properties": {
          "headerLabel": {
            "type": "string",
            "description": "Bold header to label the header row in the file preview",
            "example": "Header row"
          },
          "headerRow": {
            "type": "array",
            "description": "The header row of the data file",
            "items": {
              "type": "string",
              "example": [
                "seqID",
                "source",
                "type",
                "start",
                "end",
                "score",
                "strand",
                "phase"
              ]
            }
          },
          "snippetLabel": {
            "type": "string",
            "description": "Bold header to label the file snippet in the file preview",
            "example": "Sequence"
          },
          "fileSnippet": {
            "type": "array",
            "description": "The first interesting line of the file",
            "items": {
              "type": "string",
              "example": [
                "NC_000001.11",
                "RefSeq",
                "region  1",
                "248956422",
                ".",
                "+",
                ".",
                "ID=id0;Dbxref=taxon:9606;Name=1;chromosome=1;gbkey=Src;genome=chromosome;mol_type=genomic DNA"
              ]
            }
          }
        },
        "xml": {
          "name": "DataFilePreview"
        }
      },
      "DataFileDescriptor": {
        "type": "object",
        "description": "Array of items describing the file, e.g. row count or sequence type",
        "properties": {
          "attributeName": {
            "type": "string",
            "example": "Number of entities in the file"
          },
          "attributeValue": {
            "type": "string",
            "example": "37"
          }
        }
      },
      "DataFilePropertiesQuestion": {
        "type": "object",
        "required": [
          "questionId",
          "questionHeader",
          "possibleAnswers"
        ],
        "properties": {
          "questionId": {
            "type": "string",
            "example": "nucleotideOrProtein"
          },
          "questionHeader": {
            "type": "string",
            "description": "A header in bold above question for readability",
            "example": "Sequence Type"
          },
          "questionWording": {
            "type": "string",
            "description": "The text of the question",
            "example": "Does this file contain nucleotides or proteins?"
          },
          "possibleAnswers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataFilePropertiesAnswerOption"
            }
          }
        },
        "xml": {
          "name": "DataFilePropertiesQuestion"
        }
      },
      "DataFilePropertiesAnswerOption": {
        "type": "object",
        "required": [
          "answerId",
          "answerLabel"
        ],
        "properties": {
          "answerId": {
            "type": "string",
            "description": "Machine-readable value for this answer",
            "example": "primaryIdentifier"
          },
          "answerLabel": {
            "type": "string",
            "description": "Human readable answer to be displayed",
            "example": "Primary Identifier"
          },
          "isDefault": {
            "type": "boolean",
            "description": "TRUE if this is the default value",
            "example": "true"
          },
          "isSelected": {
            "type": "boolean",
            "description": "TRUE if the user selected this value in the wizard",
            "example": "false"
          }
        },
        "xml": {
          "name": "DataFilePropertiesAnswerOption"
        }
      },
      "DataFilePropertiesResponse": {
        "type": "object",
        "required": [
          "dataFile",
          "answers"
        ],
        "properties": {
          "dataFile": {
            "allOf": [{
              "$ref": "#/components/schemas/DataFile"
            }]
          },
          "answers": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "questionId": {
                  "type": "string",
                  "example": "nucleotideOrProtein"
                },
                "answerId": {
                  "type": "string",
                  "nullable": true,
                  "example": "nucleotide"
                }
              }
            }
          }
        },
        "xml": {
          "name": "DataFilePropertiesResponse"
        }
      },
...
    }
  }
}